Chaziz Web Explorer Codename NetChaziz ---) NetChaziz
To simplify the name. NetChaziz seems to be more catchier and even references a browser from the past, Netscape.

It should also be noted that the Alpha 1.8 Pre-Release 3 is now the Alpha 1.8 RC. It's just because it's near-final from my point-of-view.

Alpha 1.8's early development was kinda painful. During the transition of the incomplete Alpha 1.7 and the better Alpha 1.8, Multiple design changes were done.

The pre-releases showed the transition from 1.7's kinda lackluster UI to 1.8's better UI looking like a brand-new basic browser.

The RTM Alpha 1.8 should include a new STOP button.